requirements.txt
joblib==0.15.1
falcon==2.0.0
requests==2.18.4
matplotlib==3.1.3
wheel==0.30.0
numpy==1.18.1
redis==3.5.3
grpcio==1.31.0
torch==1.6.0
env==0.1.0
grpc==1.0.0
h5py==2.10.0
ipython==7.18.1
Pillow==7.2.0
protobuf==3.13.0
python_dateutil==2.8.1
scikit_learn==0.23.2
neo4j==4.1.0
